# packages/client/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reservations": {
      "get": {
        "description": "Returns a paginated list of reservations with optional filtering",
        "operationId": "ReservationController_listReservations",
        "parameters": [
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "description": "Filter by client ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "restaurantId",
            "required": false,
            "in": "query",
            "description": "Filter by restaurant ID",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "mealId",
            "required": false,
            "in": "query",
            "description": "Filter by meal type ID",
            "schema": {
              "example": 3,
              "type": "number"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Start date for filtering (YYYY-MM-DD)",
            "schema": {
              "example": "2023-12-01",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "End date for filtering (YYYY-MM-DD)",
            "schema": {
              "example": "2023-12-31",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedReservationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "List reservations",
        "tags": [
          "Reservations"
        ]
      },
      "post": {
        "description": "Creates a new reservation",
        "operationId": "ReservationController_createReservation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reservation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create reservation",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{id}": {
      "get": {
        "description": "Returns a single reservation by ID",
        "operationId": "ReservationController_getReservation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Reservation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get reservation by ID",
        "tags": [
          "Reservations"
        ]
      },
      "put": {
        "description": "Completely replaces an existing reservation",
        "operationId": "ReservationController_replaceReservation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reservation updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Reservation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Replace reservation",
        "tags": [
          "Reservations"
        ]
      },
      "patch": {
        "description": "Partially updates an existing reservation",
        "operationId": "ReservationController_updateReservation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchReservationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reservation updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Reservation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Update reservation",
        "tags": [
          "Reservations"
        ]
      },
      "delete": {
        "description": "Deletes a reservation",
        "operationId": "ReservationController_deleteReservation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Reservation deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Reservation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Delete reservation",
        "tags": [
          "Reservations"
        ]
      }
    },
...
    "/restaurants": {
      "get": {
        "description": "Returns a paginated list of restaurants in the hotel",
        "operationId": "RestaurantController_listRestaurants",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Restaurants retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRestaurantResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "List restaurants",
        "tags": [
          "Restaurants"
        ]
      }
    },
...
  },
  "info": {
    "title": "Brandon API",
    "description": "API docs for Brandon",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3040",
      "description": "Local"
    }
  ],
  "components": {
    "schemas": {
...
      "ReservationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier for the reservation",
            "example": 1
          },
          "clientId": {
            "type": "number",
            "description": "Client ID",
            "example": 1
          },
          "restaurantId": {
            "type": "number",
            "description": "Restaurant ID",
            "example": 2
          },
          "date": {
            "type": "string",
            "description": "Reservation date (YYYY-MM-DD)",
            "example": "2023-12-25"
          },
          "mealId": {
            "type": "number",
            "description": "Meal type ID (1: Breakfast, 2: Lunch, 3: Dinner, etc.)",
            "example": 3
          },
          "numberOfGuests": {
            "type": "number",
            "description": "Number of guests",
            "example": 4
          },
          "specialRequests": {
            "type": "string",
            "description": "Special requests or notes for the reservation",
            "example": "Window table preferred, celebrating anniversary"
          },
          "status": {
            "type": "string",
            "description": "Reservation status",
            "enum": [
              "confirmed",
              "pending",
              "cancelled"
            ],
            "example": "confirmed"
          }
        },
        "required": [
          "id",
          "clientId",
          "restaurantId",
          "date",
          "mealId",
          "numberOfGuests"
        ]
      },
      "PaginatedReservationResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "description": "Total number of reservations matching the criteria",
            "example": 42
          },
          "next": {
            "type": "object",
            "description": "URL to the next page",
            "example": "/api/reservations?page=2",
            "nullable": true
          },
          "previous": {
            "type": "object",
            "description": "URL to the previous page",
            "example": "/api/reservations?page=1",
            "nullable": true
          },
          "results": {
            "description": "List of reservation results",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationResponse"
            }
          }
        },
        "required": [
          "count",
          "results"
        ]
      },
      "CreateReservationRequest": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "number",
            "description": "Client ID",
            "example": 1
          },
          "restaurantId": {
            "type": "number",
            "description": "Restaurant ID",
            "example": 2
          },
          "date": {
            "type": "string",
            "description": "Reservation date (YYYY-MM-DD)",
            "example": "2023-12-25"
          },
          "mealId": {
            "type": "number",
            "description": "Meal type ID (1: Breakfast, 2: Lunch, 3: Dinner, etc.)",
            "example": 3
          },
          "numberOfGuests": {
            "type": "number",
            "description": "Number of guests",
            "example": 4,
            "minimum": 1,
            "maximum": 100
          },
          "specialRequests": {
            "type": "string",
            "description": "Special requests or notes for the reservation",
            "example": "Window table preferred, celebrating anniversary"
          }
        },
        "required": [
          "clientId",
          "restaurantId",
          "date",
          "mealId",
          "numberOfGuests"
        ]
      },
      "UpdateReservationRequest": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "number",
            "description": "Client ID",
            "example": 1
          },
          "restaurantId": {
            "type": "number",
            "description": "Restaurant ID",
            "example": 2
          },
          "date": {
            "type": "string",
            "description": "Reservation date (YYYY-MM-DD)",
            "example": "2023-12-25"
          },
          "mealId": {
            "type": "number",
            "description": "Meal type ID (1: Breakfast, 2: Lunch, 3: Dinner, etc.)",
            "example": 3
          },
          "numberOfGuests": {
            "type": "number",
            "description": "Number of guests",
            "example": 4,
            "minimum": 1,
            "maximum": 100
          },
          "specialRequests": {
            "type": "string",
            "description": "Special requests or notes for the reservation",
            "example": "Window table preferred, celebrating anniversary"
          }
        },
        "required": [
          "clientId",
          "restaurantId",
          "date",
          "mealId",
          "numberOfGuests"
        ]
      },
      "PatchReservationRequest": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "number",
            "description": "Client ID",
            "example": 1
          },
          "restaurantId": {
            "type": "number",
            "description": "Restaurant ID",
            "example": 2
          },
          "date": {
            "type": "string",
            "description": "Reservation date (YYYY-MM-DD)",
            "example": "2023-12-25"
          },
          "mealId": {
            "type": "number",
            "description": "Meal type ID (1: Breakfast, 2: Lunch, 3: Dinner, etc.)",
            "example": 3
          },
          "numberOfGuests": {
            "type": "number",
            "description": "Number of guests",
            "example": 4,
            "minimum": 1,
            "maximum": 100
          },
          "specialRequests": {
            "type": "string",
            "description": "Special requests or notes for the reservation",
            "example": "Window table preferred, celebrating anniversary"
          },
          "status": {
            "type": "string",
            "description": "Reservation status",
            "enum": [
              "confirmed",
              "pending",
              "cancelled"
            ],
            "example": "confirmed"
          },
          "updateReason": {
            "type": "string",
            "description": "Reason for update",
            "example": "Guest requested a different date"
          },
          "notifyClient": {
            "type": "boolean",
            "description": "Whether to notify the client about changes",
            "example": true
          }
        }
      },
...
      "RestaurantResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier for the restaurant",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the restaurant",
            "example": "Le Bistro Californien"
          },
          "description": {
            "type": "string",
            "description": "Description of the restaurant",
            "example": "A cozy bistro offering California-inspired French cuisine"
          },
          "capacity": {
            "type": "number",
            "description": "Maximum number of guests the restaurant can accommodate",
            "example": 80
          },
          "openingHours": {
            "type": "string",
            "description": "Restaurant opening hours",
            "example": "Mon-Sun: 12:00-15:00, 19:00-23:00"
          },
          "location": {
            "type": "string",
            "description": "Physical location of the restaurant within the hotel",
            "example": "Ground Floor, West Wing"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the restaurant is currently active",
            "example": true
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "capacity",
          "openingHours",
          "location",
          "isActive"
        ]
      },
      "PaginatedRestaurantResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "description": "Total number of restaurants matching the criteria",
            "example": 5
          },
          "next": {
            "type": "object",
            "description": "URL to the next page",
            "example": "/api/restaurants?page=2",
            "nullable": true
          },
          "previous": {
            "type": "object",
            "description": "URL to the previous page",
            "example": "/api/restaurants?page=1",
            "nullable": true
          },
          "results": {
            "description": "List of restaurant results",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RestaurantResponse"
            }
          }
        },
        "required": [
          "count",
          "results"
        ]
      },
...
    }
  }
}